Add support for multiple columns in legend
The 'legend' block now understands `columns`, and the plots now
understand `legend-columns` to set the number of columns in the
legend.

A basic flowing algorithm is used to place the entries in the legend.
The data is always the first entry in the legend. The signal entries
are also added to the first column. If there's more than one, the signal
entries follow directly the data, and the MC entries span over the remaining
columns. If there's only one, the MC entries follow directly the data, and
the signal entries are added after.