Add basic unit tests
Tests run plotIt to produce histograms, and check that the produced
pictures is identical to a reference image.

Currently, only a default plot with and without ratio are produced,
but we can easily add more test cases.

In the future, we may also want to start doing real unit-testing by
testing piece of code...