Optimize globbing operation
First, only list the content of the first file if at least one plot has
a globbing operator (*, ? or [), as most of the times there's not.

Second, if user use some globbing, the content of the files is read
once, and a string representation of the file's content is built,
without actually reading the objects inside the file. Globbing is then
performed on the representation, and not on the file's structure itself.