Merge histograms from a same group before stacking
Currently, all backgrounds are added to the stack, but assigned the same
colors and style if they belong to the same group. This cause an issue
when trying to find the minimum value of the stack, since it'd always
return the minimum value of the first background added, which may be
part of a group.

The solution is to merge histograms from a same group into a single
histogram, and adding this to the stack. Like this, the minimum is
automatically the minimum of the group and not one of its member.

Additionnally, it fixes some style issues we may encounter which made
visible the members of the group with a small line between histograms.